consolidate io into core
