+++ crates/ferritin-core/src/lib.rs
@@ -13,7 +13,8 @@
 //!
 mod atomcollection;
 mod bonds;
-mod conversions;
+mod io;
+// mod conversions;
 mod featurize;
 mod info;
 mod views;
@@ -21,3 +22,4 @@ mod views;
 pub use self::atomcollection::AtomCollection;
 pub use self::bonds::{Bond, BondOrder};
 pub use self::featurize::StructureFeatures;
+pub use self::io::load_structure;
