--- crates/ferritin-io/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "ferritin-io"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-description.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-ferritin-core = { path = "../ferritin-core" }
-memchr.workspace = true
-pdbtbx.workspace = true
-
-
-[dev-dependencies]
-ferritin-test-data = { path = "../ferritin-test-data" }
